Load Xenium OME tiff image

Load morphology OME tiff images from Xenium output with the help of RBioFormats read.image() function

LoadXeniumImage(
  path = "/Volumes/colcc-ITH2_scRNA_datasets/data/10X_Xenium_CRC_public/Breast_5K/",
  channel_cols = c("gray70", "limegreen", "cyan", "indianred3"),
  res = 5
)

Arguments

path

character string indicating the path of the Xenium OME tiff image

channel_cols

vector of color names to be assigned to each of the OME image channel

res

integer specifying resolution levels to read.

Value

Xenium image as a multi-layer SpatRaster object