hoodscanR clustering

hoodscanR clustering performed by calling the R package hoodscanR through Kandinsky

hoodscanR_clust(
  data = NULL,
  k = 100,
  label = NULL,
  n_clust = 5,
  sample_key = NULL
)

Arguments

data

Seurat object containing Kandinsky data

k

numeric, number of nearest neighbours to consider for hoodscanR clustering

label

character string specifying the variable name to be used to defne cell annotation groups

n_clust

numeric, number of expected clusters. Can be a single number or a numeric vector if multiple cluster numbers needs to be tested

sample_key

character string specifying a variable stored in the Seurat object to use as sample/batch annotation. If not NULL, neighbour networks will be defined separately for each sample/batch. Default is NULL

Value

updated Seurat object with 'hoodscanR_clusters' annotation added to metadata slot

Details

hoodscanR must be installed within the same environmnent of Kandinsky