Neighbourhood-based multiple linear regression of gene expression values

Fit a multiple linear regression for a gene of interest using neighbourhood cell count for each cell type as independent variables

nb_lm(
  seurat,
  label = NULL,
  which = NULL,
  features = NULL,
  layer = "data",
  rand = NULL,
  minprop = 0.1
)

Arguments

seurat

a Seurat object containing Kandinsky data slot

label

character string specifying the name of the variable to be used for cell type annotation.

which

name of the cell type to consider for fitting linear models. Currently only one cell type is accepted as input

features

character vector containing gene identifiers to be fitted

layer

name of Seurat layer to use for gene expression data

rand

name of random variables to be included in a mixed effect model.

minprop

numeric, genes expressed in less than this proportion of cells will be excluded from the analysis

Value

data frame reporting, for each gene, the regression coefficient and p-value associated with each cell type

Details

This function is based on fastLm() function from RcppEigen pakage. In case user wanted to include one or more confounfing (random) variables into the model, the function lmer() in instead used to fit a mixed effect model. In both cases, the output is a data frame reporting regression coefficient and pvalue for each independent variable.