Prepare Vizgen Merscope Seurat data

Initial formatting of Merscope data to work with Seurat and Kandinsky

prepare_merscope_seurat(
  path = NULL,
  dataset.id = "X",
  pattern = NULL,
  fovs = NULL
)

Arguments

path

character string specifying the path to the Merscope input files directory

dataset.id

character string that will be used to name the output Seurat object identity

pattern

character string that will be used as a key to identify Merscope input file names in case when input files from multiple Merscope samples/slides are stored in the same directory

fovs

vector of FOV identifiers to subset for downstream analysis.

Value

a Seurat object containing Merscope count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions

Details

This function will use Vizgen Merscope raw input data to build a new Seurat object. All Merscope input files must be stored in the same folder that will be specified to call the function

See also