Prepare 10X Visium HD Seurat data

Initial formatting of Visium data to work with Seurat and Kandinsky

prepare_visiumHD_seurat(
  path = NULL,
  dataset.id = "X",
  img.res = c("full", "high", "low"),
  fullres_path = NULL,
  binsize = c(2, 8, 16)
)

Arguments

path

character string specifying the path to the Visium input files directory

dataset.id

character string that will be used to name the output Seurat object identity

img.res

character string specifying which version of H&E Visium image will be loaded into the Seurat data. Must be one of the following: `low`, `high`, `full`, where `full` refers to the original H&E full-resolution tiff image. If set to `full`, lowres image will be loaded into Seurat, but the path to the full res tif image will be also stored in the final object.

fullres_path

character string specifying the path to the Visium full resolution H&E tif image. The full image path will be stored into the final Seurat object when `img.res` is set to `full`, and will be used by `Kandinsky` functions for downstream analyses

binsize

numeric vector specifying the binsize(s) to be used to build the Seurat object. Can be any combination of `2`, `8`, and `16` sizes

Value

a Seurat object containing Visium count matrix, metadata, and additional files/file paths to be used for downstream analysis with Kandinsky functions

Details

This function will use 10X Visium raw input data to build a new Seurat object. All Visium input files must be stored in the same folder that will be specified to call the function

See also