Kandinsky S4 object class

Kandinsky is a list object storing spatial data and information for spatial transcriptomic dataset analysed through Seurat.

Slots

platform

sequencing technology

nb

weighted neighbour list object created with the package`spdep`

sf

geometry data frame created with the package `sf` and containing cell/spot coordinates

nb.type

character indicating the neighbouring strategy used to initialize Kandinsky object. Must be one of the following: `Q`, `K`, `C`, `M`

nnMat

list object containing nearest neighbour matrix created with one of the `nnMat` functions implemented in Kandinsky

spot_distance

numeric value indicating centroid-to-centroid distance between Visium spots or Visium HD bins

texture

data.frame containing GLCM texture features extracted from Visium H&E slide image

tx

character string indicating the path for the transcript coordinates file from CoxMx/Xenium/Merscope platforms

fov_mask

geometry data frame created with the package `sf` and containing CosMx FOVs bounding box coordinates

img

Visium H&E or any other tissue slide image stored as a SpatRaster or SpatrasterCollection object via the package `terra`

mask

masked H&E image obtained via the Kandinsky function `he_mask` and stored as a SpatRaster or SpatrasterCollection object via the package `terra`